+++ AutomatedTesting/Gem/PythonTests/Physics/CMakeLists.txt
@@ -8,6 +8,18 @@
 
 if(PAL_TRAIT_BUILD_TESTS_SUPPORTED AND PAL_TRAIT_BUILD_HOST_TOOLS)
     ly_add_pytest(
+        NAME AutomatedTesting::PhysicsTests_Main
+        TEST_SUITE main
+        TEST_SERIAL
+        PATH ${CMAKE_CURRENT_LIST_DIR}/TestSuite_Main.py
+        RUNTIME_DEPENDENCIES
+            Legacy::Editor
+            AZ::AssetProcessor
+            AutomatedTesting.Assets
+        COMPONENT
+            Physics
+    )
+    ly_add_pytest(
         NAME AutomatedTesting::PhysicsTests_Periodic
         TEST_SUITE periodic
         TEST_SERIAL
